+++ src/utility/animations/global.js
